+++ extensions/hammerspoon/CHANGELOG.md
@@ -1,3 +1,8 @@
 # Hammerspoon Changelog
 
+## [Improvements] 1.1.0 - {PR_MERGE_DATE}
+
+- add Open Documentation command
+- add option to not open console on reload
+
 ## [Initial Version] 1.0.0 - 2025-05-21